requirements.txt
openai>=1.0.0
langchain>=0.3.0
langgraph>=0.2.0
pyyaml>=6.0
jinja2>=3.1.0
rich>=13.0.0
click>=8.1.0
pydantic>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
